refactor: internal/repo — split walk.go → ignore.go (v0.8.1 Part A)
Move the minimal gitignore-rule plumbing (rule, gitignore,
scopedGitignore, loadGitignoreFS, parseGitignore, pruneScopes,
matchScopes, relToScope, collectGitignores, compileRule) from
walk.go into a sibling ignore.go in the same package. walk.go
retains the WalkFS + Walk + Matcher + ShouldIndex + isBinary
machinery — the file-acquisition surface — while ignore.go
isolates the .gitignore engine the scope stack walks against.

Zero behavior change. Same package boundary; no caller changes.
The "regexp" import drops out of walk.go since the rule compiler
lives in ignore.go now.
